move request data to agg request files
